# Conflicts:
#	Premiere/EncodingData.h

+++ Premiere/EncodingData.h
@@ -9,17 +9,7 @@ typedef struct EncodingData
 	int planes;
 	const char *pix_fmt;
 	char *plane[8] = {};
-	unsigned int stride[8] = 
-	{ 
-		static_cast<unsigned>(-1),
-		static_cast<unsigned>(-1),
-		static_cast<unsigned>(-1),
-		static_cast<unsigned>(-1),
-		static_cast<unsigned>(-1),
-		static_cast<unsigned>(-1),
-		static_cast<unsigned>(-1),
-		static_cast<unsigned>(-1)
-	};
+	int32_t stride[8] = { -1, -1, -1, -1, -1, -1, -1, -1 };
 	int pass;
 	struct filters
 	{
